# cmp: Compare request times