Before:
```
λ warp cmp before.csv.zst after.csv.zst
274918 operations loaded... Done!
366324 operations loaded... Done!
-------------------
Operation: DELETE
Operations: 27495 -> 36629
* Average: +33.21% (+30.4) obj/s
-------------------
Operation: GET
Operations: 123710 -> 164845
* Average: +33.25% (+0.1 MiB/s) throughput, +33.25% (+137.1) obj/s
* First Byte: Average: -3.951011ms (-17%), Median: -757.1µs (-27%), Best: +1.8µs (+0%), Worst: -506.2678ms (-16%)
-------------------
Operation: PUT
Operations: 41237 -> 54949
* Average: +33.26% (+0.0 MiB/s) throughput, +33.26% (+45.7) obj/s
-------------------
Operation: STAT
Operations: 82476 -> 109901
* Average: +33.26% (+91.4) obj/s
```

After:
```
λ warp cmp before.csv.zst after.csv.zst
274918 operations loaded... Done!
366324 operations loaded... Done!
-------------------
Operation: DELETE
Operations: 27495 -> 36629
* Average: +33.21% (+30.4) obj/s
* Requests: Avg: -24.6ms (-35%), P50: -1.9756ms (-28%), P99: -330.2085ms (-26%), Best: -56.6µs (-3%), Worst: -628.0938ms (-7%)
-------------------
Operation: GET
Operations: 123710 -> 164845
* Average: +33.25% (+0.1 MiB/s) throughput, +33.25% (+137.1) obj/s
* Requests: Avg: -3.95ms (-17%), P50: -795.3µs (-28%), P99: -54.9509ms (-7%), Best: -500ns (-0%), Worst: -506.2678ms (-16%)
* TTFB: Avg: -3.95ms (-17%), P50: -757.1µs (-27%), P99: -54.9509ms (-7%), Best: +1.8µs (+0%), Worst: -506.2678ms (-16%)
-------------------
Operation: PUT
Operations: 41237 -> 54949
* Average: +33.26% (+0.0 MiB/s) throughput, +33.26% (+45.7) obj/s
* Requests: Avg: -23.7ms (-33%), P50: -10.3µs (-0%), P99: -317.7879ms (-26%), Best: -301.8µs (-10%), Worst: +716.2566ms (+9%)
-------------------
Operation: STAT
Operations: 82476 -> 109901
* Average: +33.26% (+91.4) obj/s
* Requests: Avg: -3.1ms (-14%), P50: -216µs (-10%), P99: -24.2622ms (-3%), Best: -2.1µs (-0%), Worst: -1.7413943s (-44%)
```